+++ JediCodeplug/ExternalBlocks/Block53.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
@@ -12,11 +13,8 @@ namespace JediCodeplug
         private byte[] _contents;
         public Span<byte> Contents { get => _contents; set => _contents = value.ToArray(); }
 
-        public override int Id { get => 0x53; }
+        public override byte Id { get => 0x53; }
         public override string Description { get => "Scan List"; }
-
-        #region Propeties
-        #endregion
 
         #region Definition
         /*  0  1  2  3   4  5  6  7    8  9  A  B   C  D  E  F
@@ -27,6 +25,9 @@ namespace JediCodeplug
 
         #endregion
 
+        #region Propeties
+
+        #endregion
 
         public Block53() { }
 
@@ -39,7 +40,7 @@ namespace JediCodeplug
         {
             var contents = Contents.ToArray().AsSpan(); //TODO
             var nextAddress = address + Contents.Length + BlockSizeAdjustment;
-            
+
             Serializer(codeplugContents, address, contents);
             return nextAddress;
         }
