+++ JediCodeplug/InternalBlocks/Block06.cs
@@ -5,7 +5,7 @@ namespace JediCodeplug
 {
     public class Block06 : Block
     {
-        public override int Id { get => 0x06; }
+        public override byte Id { get => 0x06; }
         public override string Description { get => "Softpot Vector"; }
 
         #region Definition
